stop using pid_t in a bizarre and brain-addled way
Salvador Fandiño reported, with a charitable politeness, that I had
somehow been extremely disturbed when writing this code.  I used
pid_t repeatedly, as if the integer status gotten by waitpid was a
pid_t when of course it is not.  It's just an integer.

(I must have just remembered that waitpid returns a pid_t and not
engaged my brain at all.)

Ruby's Process::Status provides this in a `.to_i` method, which is
not as appropriate in Perl, so I have taken Salvador Fandiño's
suggestion and called this datum the return_code.